requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
PyJWT>=2.8
qrcode
cryptography
